Clean up dark mode error visibility for reset password (#1236)
* Clean up dark mode error visibility for reset password

* Relocates error message container

Moves the error message container to be positioned directly after the email input field. This improves the user experience by providing immediate feedback in close proximity to the input field.

* Refactors email input error display

Simplifies the email input error display by removing the dedicated
errorMessage target and leveraging the existing viral_help_text
component for rendering error messages.

Removes blur event listener, as input validation is handled on input.

* Refactors email validation for reusability

Consolidates email validation logic into a single `validateEmail` function with configurable error display and focus behavior.
This eliminates the need for separate `validateEmailSilently` function and improves code maintainability.

Also, removes unnecessary binding in constructor and disconnect methods.

* Removes unused event parameter

The event parameter in the `handleInput` function was not being used,
so it is removed to simplify the code.

* Removes default values for email validation messages

Removes default values for email validation messages. This allows for greater flexibility in configuring these messages, as the defaults are no longer hardcoded in the controller.

* Validates required Stimulus values

Ensures that the `email-missing` and `email-format` Stimulus values
are provided when the controller is initialized.

Raises an error if either value is missing, preventing unexpected
behavior due to undefined values.