Adds aria-label to tabs panel
Adds an `aria-label` attribute to the tabs panel to improve accessibility.

Moves the logic for setting the `aria-label` from the Ruby component to the ERB template.
This change allows the label to be correctly rendered and associated with the tabs panel element,
enhancing the user experience for assistive technologies.

Updates system test to reflect updated html.