Improves tab panel accessibility (#1247)
* Fixes focus state

* Improves tab panel accessibility and structure

Enhances the tab panel component with ARIA attributes for better accessibility.

This commit introduces the following changes:

- Wraps the tab list in a <nav> element for improved semantic structure and accessibility.
- Adds ARIA roles (tablist, tab, presentation) and attributes (aria-selected, aria-controls, aria-labelledby) to the tab panel and tab components for enhanced accessibility.
- Implements automatic ID generation for tabs to ensure unique identifiers for ARIA attributes.
- Modifies the base tag of the component to a div.

* Reverted files

* Swaps `div` and `nav` tags for tabs panel

Updates the tabs panel component to use a `div` as the root element
instead of `nav`, and moves the `nav` tag inside.

This change aligns the HTML structure with the intended semantic
meaning and styling of the tabs panel.

* Replaces <nav> with <div> in tabs panel

Updates the tabs panel component to use a <div> element instead of a <nav> element for its container.

This change is likely to provide greater flexibility in styling and layout, while maintaining the intended functionality.

* Fixed position of neav element

* Simplifies workflow execution tab navigation

Removes redundant `within` blocks in system tests for workflow
execution tabs, streamlining the navigation and making the tests
more readable and maintainable.

Removes unnecessary selector assertions in the tabs panel component test.

* Adds aria-label to tabs panel

Adds an `aria-label` attribute to the tabs panel to improve accessibility.

Moves the logic for setting the `aria-label` from the Ruby component to the ERB template.
This change allows the label to be correctly rendered and associated with the tabs panel element,
enhancing the user experience for assistive technologies.

Updates system test to reflect updated html.

* Adds ID to tabs panel for easier testing

Adds an ID to the tabs panel component to allow for more robust
system testing, particularly when interacting with tabs. The focus
ring style is also updated to remove an unnecessary selector,
simplifying the styling.

* Removes redundant ID assignment

Removes the explicit assignment of the ID to the system arguments as it is already handled elsewhere. This simplifies the code.

* Restructures tabs panel for improved semantics

Refactors the tabs panel component to use a `` element as its root for better semantic structure.

This change also ensures the `id` attribute is consistently applied to the root element.
Additionally, the `aria-label` is dynamically added to the `` element if a label is provided, enhancing accessibility.

* Updates test selector for tabs panel

Updates the test selector in the tabs panel test to use the ID attribute instead of the class attribute.

This change improves the specificity and reliability of the test.