Table PUIDs (#1255)
* Improves table layout and styling

Enhances table component to better handle column widths, especially for the PUID column.

Introduces a helper function to calculate the PUID column width based on the PUID length and checkbox presence, ensuring consistent spacing and alignment.

Updates the table component to use CSS variables for width calculations, improving responsiveness and maintainability.

* Improves attachment table layout and display

Enhances the attachment table's layout, specifically addressing the sticky positioning and width of the first two columns. It introduces a CSS variable to manage the width, ensuring consistent display of the `puid` column. Also, wraps the `puid` with a monospace font for better readability.

* Improves UUID column width handling in tables

Ensures consistent UUID column width across workflow execution tables.

Uses a CSS variable to define the UUID column width based on
the UUID length, spacing, and the presence of a checkbox.

This provides a more robust and maintainable solution for
handling variable column widths.

* Enhances sample metadata table layout

Adjusts the width of the checkbox column in the sample metadata table for improved alignment and consistency. This change ensures better visual presentation and usability when updating samples.

* chore: fix fixed column width and offset to use same font-family for calculation and include cell padding in calculation

* chore: re-add smaller text to samples table header

* Corrects colspan calculation in table component

Corrects the colspan calculation for the sticky header and data row in the table component.

The change ensures the table header and data rows properly span the correct number of columns, accounting for removed status column

* Enhances table styling and readability

Improves the visual appearance and user experience of the attachments table.

Applies consistent font styles to sticky columns for enhanced readability.
Adds a flex layout to filename column

* Updates attachment table PUID width calculation

Updates the attachment table to dynamically adjust the PUID width based on whether the user has the ability to select attachments. This ensures the table layout adapts correctly to the presence of checkboxes for attachment selection.

* Refactors CSS width calculations for IDs

Improves the logic for calculating the width of PUID and UUID columns.

Centralizes width calculation into a private helper, improving consistency and reducing code duplication.

Adjusts spacing multipliers and adds fixed pixel offsets for checkbox alignment, enhancing visual consistency.

---------

Co-authored-by: Eric Enns <492127+ericenns@users.noreply.github.com>