Adds refresh notice component
Implements a refresh notice component to inform users when data is out of date and provide a button to refresh.

The component utilizes Turbo Streams for real-time updates and includes logic to prevent redundant refreshes triggered by editable cells. It also uses a stimulus controller to handle dismissing the notice and reloading the page. The component is now rendered in the samples index view.