A11Y Sidebar Improvements (#1282)
* Update navbar localization keys for consistency in layout components

* Refactor layout components to use semantic HTML elements

- Change `<div>` to `<header>` in `layout_component.html.erb` for better structure.
- Update `<aside>` to `<nav>` in `sidebar_component.html.erb` to reflect navigation semantics.
- Replace `<div>` with `<header>` in `header_component.html.erb` for improved accessibility.
- Remove unused localization key `navigation` from `en.yml` and `fr.yml` for consistency.

* Refactor flash message container for improved readability

- Adjust the formatting of the `flashes` div in `layout_component.html.erb` for better code clarity and maintainability.

* Update navbar button localization keys in system tests

* Update sidebar selector from 'aside#sidebar' to 'nav#sidebar' in system tests

* Refactor sidebar component tests to use 'nav' instead of 'aside' and improve accessibility checks

* Update sidebar tests to assert 'nav#sidebar' instead of 'aside' for improved semantic accuracy

* Uses consistent key for new dropdown label

Updates the sidebar "new" dropdown button title and aria label to use the
`general.navbar.new_dropdown.label` key for consistency and to avoid
using the generic `common.actions.create` key.
This ensures the button text accurately reflects its function.

* feat: add announcements for sidebar collapse and expand states in layout component

* Adds aria-expanded attributes to sidebar toggle

Improves accessibility by adding aria-expanded attributes to the sidebar toggle buttons.

This allows screen readers to correctly announce the current state of the sidebar (collapsed or expanded). Focus is also returned to the toggle button that was interacted with to improve the user experience.

* Remove default announcementsEnabled setting in layout controller

* Updates sidebar toggle labels for accessibility

Updates the aria labels and titles for the sidebar toggle buttons to use a more descriptive "Toggle sidebar menu" label.

This improves accessibility by providing clearer instructions to screen reader users, making it easier for them to understand the function of the buttons.

Removes the specific collapse/expand labels as the single toggle now handles both actions.

* Update sidebar component tests to use localized aria labels for improved accessibility