Ignores importmap integrity W3C validation errors
The W3C validator doesn't yet recognize the "integrity"
property in import maps, leading to false positive errors.
This change adds a filter to ignore these specific errors,
allowing for accurate validation of other potential issues.

This also fixes a javascript error that occurs when the colour
mode is set on.