Fix: Colour theme selection announcement error + Import map enable_integrity (#1341)
* Fixes colour mode controller error

* Add asset integrity hash algorithm configuration and enable integrity for importmap pins

* Ignores importmap integrity W3C validation errors

The W3C validator doesn't yet recognize the "integrity"
property in import maps, leading to false positive errors.
This change adds a filter to ignore these specific errors,
allowing for accurate validation of other potential issues.

This also fixes a javascript error that occurs when the colour
mode is set on.

* Ignores aria-label errors on divs without role

Addresses W3C validation errors related to the use of `aria-label` attributes on `div` elements lacking appropriate ARIA roles.

The W3C Nu HTML Checker now enforces that `aria-label` should only be used on `div` elements when they have a valid role. This change adds a filter to ignore these specific errors, preventing false positives during validation.

This is a workaround to allow the colour-mode.js related changes to be merged, as the javascript was generating these validation errors in CI and blocking deployment.

* Fixes ARIA describedby attribute for scopes

Updates the ARIA describedby attribute for the scope checkboxes to correctly handle both help text and error messages.

This ensures that assistive technologies can properly convey the purpose and validation status of each scope option to the user.

* Updates validator image to latest

Updates the validator image tag in the CI workflow to 'latest'.

This ensures the workflow always uses the most recent version of the validator,
potentially including bug fixes and improvements.