Dynamically set the number of metadata fields visible on the samples table (#1349)
* Enhance Samples Table Component with Metadata Field Limit and Warning

- Introduced a maximum limit for metadata fields displayed in the samples table to improve performance.
- Added a warning message when the number of metadata fields exceeds the defined limit.
- Updated the HTML structure for better readability and consistency.
- Made minor adjustments to localization files for improved string formatting.

* Improves metadata template validation and messages

Enhances metadata template validation by formatting JSON errors to provide more user-friendly messages.

Updates the JSON schema to limit the number of items in the metadata template to 150.

Refactors locale files to provide the new JSON error messages.

fix/limit-max-metadata-fields-samples-table

* Normalized files

* Uses I18n for metadata template error messages

Updates the metadata template update service test to use I18n for
error messages, ensuring consistency and support for localization.

The error messages related to invalid fields were previously hardcoded,
which is not ideal for maintainability and internationalization. This
change replaces the hardcoded strings with I18n lookups.

* Updates metadata template creation tests

Updates the metadata template creation tests to use I18n for error message assertions.
This change ensures that the tests are more robust and less prone to failure due to changes in the default error messages.

* Dynamically limits metadata fields in sample table

Limits the number of metadata fields displayed in the sample table based on the number of samples to improve performance and prevent excessive table size.

Calculates the maximum number of metadata fields to display based on the target number of cells (rows * columns), capping the maximum fields at 200.

Updates the warning message to reflect the dynamic limit and provides more context to the user.

* Limits metadata fields displayed in samples table

Improves performance by limiting the number of metadata fields displayed in the samples table, especially when there are many samples.

It calculates the maximum number of metadata fields based on the sample count to keep the number of table cells within a target range. A warning message is displayed if the number of fields is limited, with a link to create a metadata template.

Refactors the logic for applying the metadata field limit into a dedicated method. Also includes fix for Metadata Templates create/update error messages.

* Adds validation for max fields in metadata template

Adds a validation to limit the number of fields in a metadata
template to a maximum of 150. This prevents potential performance
issues and maintains data integrity when rendering samples table.

* Moves URL helpers to a dedicated module

Refactors the Samples::TableComponent to improve code organization by extracting URL generation logic into a dedicated UrlHelpers module.

This change enhances the component's readability and maintainability by separating concerns, making it easier to understand and modify the URL-related functionality.

* Removed unwanted inspection code

* Updates metadata template URL

Updates the URL for creating metadata templates to use the index path instead of the new path.

This ensures consistency and aligns with the intended navigation flow for managing metadata templates.

* Normalized translations

* Reverted css class chagne

* Refactors metadata fields size warning message

Refactors the logic for displaying the metadata fields size warning message in the samples table component.

Moves the logic for building the warning message, including the conditional link to create a metadata template, into dedicated methods within the component. This improves code readability and maintainability by separating concerns. The template link is now only rendered if the user has edit permissions.

Updates locale files to support separate messages for users with and without edit permissions.

* Ignore rubocop error